requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Django==5.2.6
djangorestframework==3.16.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
redis==6.4.0
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
zope.interface==7.2
